Several Improvements to Fitting + Move from Commons Math 3.5 to 3.6.1
Function fitting has been improved in several ways. Obtaining an initial
point is now much easier: We use the same optimization procedure for
this as for solving the real problem, just based on n points only, where
n is the number of parameters. If there are n parameters and n points,
it should theoretically be possible to fit a function perfectly to them,
unless the function cannot be made to fit or some of the points have the
same x-value. Anyway, this way we have moved much of the complexity of
getting good start points from the model into the fitting procedure.
This should make adding new models much easier. We can also use fewer
starting points.

Additionally, we now also use an "afterburner" optimization step: Once
the normal procedure is finished, we apply a BOBYQA step. This should
help to squeeze the last ounce of precision out of the fitting.