Fixed Some Bugs: _LSSimplexFittingJob now worse :-(
I fixed a couple of issues, most notably added a work-around for issue
https://issues.apache.org/jira/browse/MATH-1376
of SimplexOptimizer. This issue would lead the optimizer to die with an
exception if it could not reach convergence before the predefined
number of steps was exhausted. I want it to simply return the best
solution it had so far instead.

I therefore added an internal convergence checker and an additional
counter. Now it should normally terminate gracefully. This leads
_LSSimplexFittingJob to become slower and return worse results :-(