Improved Structure of MultiFittingJob + Other Minor Improvements
MultiFittingJob will now not necessarily throw an error if one of its
sub-jobs fails: instead, as one of them succeeds, it will return the
discovered model. This is to deal with situations where some it is
impossible to fit a specific model to a given data set while another one
from the model collection does fit.

The mathPrint for ExponentialDecayModel has further been improved.

ClusteringExample and FittingExample have been adapted to the changes in
Execute.