QuadraticModel Parameter Guesser Adapted to Changes in Polynomial
The parameter guesser of QuadraticModel has been adapted to the changes
in the Polynomial class of utils-math. The coefficient computation
routines in there now return the fitting error instead of true or false.
Also, the coefficient getter for quadratic functions in utils-math has
been improved to additionally attempt to get the coefficients by solving
a system of linear equations. This solution process, in turn, applies
several different decomposition strategies from Apache Commons Math as
well as Gaussian elimination to obtain the best (lowest-error)
solutions for linear equations. All in all, if we have three points, we
can compute the parameters of the corresponding quadratic function very
precisely.