Added Evolution Strategy + Least-Squares-based Function Fitter
A function fitter based on a (mu+lambda)-Evolution Strategy which
refines its population every now and then with a Least-Squares approach
(Levenberg-Marquardt) has been added. This fitter seems to perform well
on several of the test problems.

Our standard fitting arsenal is now composed of
1. (mu+lambda)-Evolution Strategy + Least Squares
2. Differential Evolution + Least Squares
3. CMA-ES + Least Squares
4. Least Squares + Simplex + BOBYQA

Using several of these algorithms in combination should help us to get
stable and robust results. We can expect that for each one of these
algorithms, there are scenarios where it performs best. This seems to
be indicated by our FittingExamples test.

By applying all of these algorithms to a given fitting task via a
MultiFittingJob, we probably have a good chance to find a fitting close
to the global optimum for a given model (and the closest-to-optimal
model).